+++ api/views/employee_views.py
@@ -10,7 +10,6 @@ from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework.exceptions import PermissionDenied
 from rest_framework.permissions import AllowAny, IsAuthenticated, IsAuthenticatedOrReadOnly
-from oauth2_provider.contrib.rest_framework import TokenHasReadWriteScope, TokenHasScope
 from api.pagination import CustomPagination
 from django.db.models import Q
